feat(core): drop disallowed marks instead of deleting plain blocks
When a block's content type changes from "inline" to "plain" (e.g. code
blocks), old documents carry formatting marks the plain node no longer
allows. Previously these caused data loss: y-prosemirror deletes the whole
block on reconstruction, and BlockNote-JSON rehydration threw on createChecked.

Instead of mutating the source (the removed stripDisallowedMarks pre-sync
rule), leave the marks in place and never materialize the disallowed ones:

- Yjs path: the FixUpSchema extension (formerly FixupCreateAndFill, renamed
  since it now does more than fix createAndFill) wraps schema.node to filter
  each text child's marks through NodeType.allowedMarks during reconstruction.
- Write path: styledTextToNodes drops disallowed marks for plain blocks
  before building the node, fixing initialContent / insert / replace /
  updateBlock / exporters.

Both are scoped to plain-content blocks via the shared isPlainContentNodeType
helper, so inline blocks keep their formatting. The read path already strips
via textContent. Comment/suggestion (annotation) marks stay allowed and are
preserved.
