+++ packages/core/src/yjs/extensions/schemaMigration/SchemaMigration.ts
@@ -1,4 +1,3 @@
-import { Schema } from "@tiptap/pm/model";
 import { Plugin, PluginKey } from "@tiptap/pm/state";
 import * as Y from "yjs";
 
@@ -6,80 +5,24 @@ import {
   createExtension,
   ExtensionOptions,
 } from "../../../editor/BlockNoteExtension.js";
-import migrationRules, {
-  preSyncMigrationRules,
-} from "./migrationRules/index.js";
+import migrationRules from "./migrationRules/index.js";
 
 // This plugin allows us to update collaboration YDocs whenever BlockNote's
-// underlying ProseMirror schema changes. There are two kinds of migration:
-//
-// - Pre-sync (`preSyncMigrationRules`): mutate the Yjs fragment directly and
-//   must run BEFORE y-prosemirror reconstructs the document. This is required
-//   when invalid content would make y-prosemirror reject a node — its error
-//   handler DELETES the whole node from the Yjs doc (propagating the deletion
-//   to all peers), so the fragment has to be fixed first.
-// - Post-sync (`migrationRules`): repair the reconstructed ProseMirror document
-//   via an appended transaction, for changes where the node survives sync.
+// underlying ProseMirror schema changes. The plugin reads the current Yjs
+// fragment and dispatches additional transactions to the ProseMirror state, in
+// case things are found in the fragment that don't adhere to the editor schema
+// and need to be fixed. These fixes are defined as `MigrationRule`s within the
+// `migrationRules` directory.
 export const SchemaMigration = createExtension(
-  ({ editor, options }: ExtensionOptions<{ fragment: Y.XmlFragment }>) => {
-    const { fragment } = options;
-
-    // Runs the pre-sync rules over the whole fragment. Safe to call repeatedly:
-    // the rules are no-ops once the fragment is clean, so the transaction they
-    // run in produces no change (and thus doesn't re-trigger the observer). The
-    // rules need the ProseMirror schema, which only exists once the editor has
-    // been constructed — so callers pass it in.
-    const runPreSyncMigrations = (schema: Schema) => {
-      fragment.doc?.transact(() => {
-        for (const rule of preSyncMigrationRules) {
-          rule(fragment, schema);
-        }
-      });
-    };
-    // The observer is registered before the editor mounts (so it precedes
-    // y-prosemirror's), but only fires later — by which point `editor.pmSchema`
-    // is available.
-    const preSyncObserver = () => {
-      if (editor.pmSchema) {
-        runPreSyncMigrations(editor.pmSchema);
-      }
-    };
-    // `unobserveDeep` is a no-op if the observer isn't (or is no longer)
-    // registered, so this is safe to call more than once.
-    const stopPreSyncObserver = () => fragment.unobserveDeep(preSyncObserver);
-
-    // Migrate content that streams in from a provider after mount. Registered
-    // here — in the extension factory, before the editor mounts — so it runs
-    // before y-prosemirror's own observer and can clean the fragment first.
-    // Removed once migration is done (see `appendTransaction`).
-    fragment.observeDeep(preSyncObserver);
-
+  ({ options }: ExtensionOptions<{ fragment: Y.XmlFragment }>) => {
     let migrationDone = false;
     const pluginKey = new PluginKey("schemaMigration");
 
     return {
       key: "schemaMigration",
-      // Run before y-sync so the pre-sync migration (in this plugin's `view`)
-      // cleans the fragment before y-prosemirror reconstructs the document.
-      runsBefore: ["ySync"],
       prosemirrorPlugins: [
         new Plugin({
           key: pluginKey,
-          view: (view) => {
-            // Migrate content already present before mount (e.g. a document
-            // loaded from a database). `observeDeep` only fires on later
-            // changes, so existing content is handled explicitly here — in the
-            // plugin's view init, where the schema is available, and before
-            // y-prosemirror reconstructs the document.
-            if (fragment.firstChild) {
-              runPreSyncMigrations(view.state.schema);
-            }
-            return {
-              // Safety net: stop observing if the editor is destroyed before any
-              // content ever synced in (so `migrationDone` was never reached).
-              destroy: stopPreSyncObserver,
-            };
-          },
           appendTransaction: (transactions, _oldState, newState) => {
             if (migrationDone) {
               return undefined;
@@ -102,14 +45,6 @@ export const SchemaMigration = createExtension(
             }
 
             migrationDone = true;
-            // The initial document has now synced in and been migrated by the
-            // pre-sync observer (which runs before this, during the y-sync). New
-            // edits can't introduce content the schema rejects, so the observer
-            // is no longer needed — stop it instead of running on every later
-            // transaction. (Like the post-sync migration, this is one-shot:
-            // disallowed content arriving from an old-version peer *after* the
-            // first sync would not be re-migrated.)
-            stopPreSyncObserver();
 
             if (!tr.docChanged) {
               return undefined;
